requirements.txt
# ResearchPro Requirements
# Python 3.10+ required

# Core Dependencies
google-adk>=0.1.0
google-generativeai>=0.8.0
google-cloud-aiplatform>=1.60.0

# Web Tools
requests>=2.32.0
beautifulsoup4>=4.12.0
aiohttp>=3.10.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# NLP and Text Processing
nltk>=3.8.1
spacy>=3.7.0

# API and Networking
httpx>=0.27.0
urllib3>=2.2.0

# Async Support
asyncio>=3.4.3
aiofiles>=24.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.8.0
typing-extensions>=4.12.0

# Observability
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp>=1.25.0

# Testing and Evaluation
pytest>=8.3.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0

# Code Quality
black>=24.4.0
pylint>=3.2.0
mypy>=1.11.0

# Documentation
mkdocs>=1.6.0
mkdocs-material>=9.5.0

# Optional: For production deployment
# google-cloud-logging>=3.10.0
# google-cloud-trace>=1.13.0
# google-cloud-monitoring>=2.21.0
